requirements.txt
# Core dependencies
numpy>=2.0.2
pandas>=2.0.0
python-dotenv>=1.0.0

# OpenAI and LangChain
openai>=1.0.0,<2.0.0
langchain>=0.0.300,<0.1.0
langchain-openai>=0.0.5,<0.1.0
langchain-core>=0.1.0,<0.2.0
langchain-text-splitters>=0.0.1
langgraph>=0.0.1

# RAG Dependencies
faiss-cpu>=1.7.4  # or faiss-gpu if using GPU
sentence-transformers>=2.2.2
pdfplumber>=0.10.3

# Web framework
streamlit>=1.27.0,<2.0.0

# Data visualization
plotly>=5.15.0,<6.0.0

# Additional utilities
requests>=2.28.0
tiktoken>=0.5.0
PyYAML>=6.0.0
httpx>=0.25.0
anyio>=3.7.0
pydantic>=2.0.0

# For development
ipykernel>=6.0.0
ipython>=8.0.0